Extracted header into its own component Refactored index page to be more of a placeholder
