+++ src/lib/components/Nav.svelte
@@ -4,8 +4,8 @@
 
 <div class="filter drop-shadow-md bg-white-100">
 	<nav class="border-t-2 border-solid border-red-100">
-		<a href="{base}/"
-			><img src={`${base}/tplogo-red.svg`} alt="Telstra Purple" class="px-48 py-4" /></a
-		>
+		<a href="{base}/">
+			<img src={`${base}/tplogo-red.svg`} alt="Telstra Purple" class="px-48 py-4" />
+		</a>
 	</nav>
 </div>
